AppFeatureModel - Arquivo renomeado para navigator
